feat: add scopes to RegisterPublicKeyRequest
This will allow registering a public key with a specific (e.g. reduced) set of scopes.
We also add a new flag "skip_user_scopes" to the RegisterPublicKeyRequest. If it is set to true and the scopes on the request are empty, this flag will prevent the user scopes from being set on the public key as the default behavior.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>